refactor: Provider 역할 축소
